refactor: refactor execution tests
